refactor(auth): add JwtHelper, create Authentication folder, logout user on expired token
