+++ web/Layout/TopBar.razor
@@ -1,5 +1,6 @@
+@using AgendaUni.Web.Services.Authentication
 @using Microsoft.AspNetCore.Components.Authorization
-@inject AgendaUni.Web.Services.AuthService AuthService
+@inject AuthService AuthService
 
 <div class="top-row px-4 d-flex justify-content-end align-items-center">
     <AuthorizeView>
