+++ web/Pages/Login.razor
@@ -1,9 +1,10 @@
 @page "/login"
-@inject AgendaUni.Web.Services.AuthService AuthService
+@inject AuthService AuthService
 @inject NavigationManager NavigationManager
 @using System.ComponentModel.DataAnnotations
-@using AgendaUni.Web.Models
 @using System.Web
+@using AgendaUni.Web.Models
+@using AgendaUni.Web.Services.Authentication
 
 <PageTitle>AgendaUni</PageTitle>
