--- web/Services/CustomAuthenticationStateProvider.cs
@@ -1,110 +1,0 @@
-using Blazored.LocalStorage;
-using Microsoft.AspNetCore.Components.Authorization;
-using System.Security.Claims;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Threading.Tasks;
-using System.Text.Json;
-
-namespace AgendaUni.Web.Services
-{
-    public class CustomAuthenticationStateProvider : AuthenticationStateProvider
-    {
-        private readonly ILocalStorageService _localStorage;
-        private readonly HttpClient _httpClient;
-
-        public CustomAuthenticationStateProvider(ILocalStorageService localStorage, HttpClient httpClient)
-        {
-            _localStorage = localStorage;
-            _httpClient = httpClient;
-        }
-
-        public override async Task<AuthenticationState> GetAuthenticationStateAsync()
-        {
-            var authToken = await _localStorage.GetItemAsStringAsync("authToken");
-
-            var identity = new ClaimsIdentity();
-            _httpClient.DefaultRequestHeaders.Authorization = null;
-
-            if (!string.IsNullOrEmpty(authToken))
-            {
-                try
-                {
-                    identity = new ClaimsIdentity(ParseClaimsFromJwt(authToken), "jwt");
-                    _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", authToken);
-                }
-                catch
-                {
-                    await _localStorage.RemoveItemAsync("authToken");
-                    identity = new ClaimsIdentity();
-                }
-            }
-
-            var user = new ClaimsPrincipal(identity);
-            return new AuthenticationState(user);
-        }
-
-        public void MarkUserAsAuthenticated(string token)
-        {
-            var authenticatedUser = new ClaimsPrincipal(new ClaimsIdentity(ParseClaimsFromJwt(token), "jwt"));
-            var authState = Task.FromResult(new AuthenticationState(authenticatedUser));
-            NotifyAuthenticationStateChanged(authState);
-        }
-
-        public void MarkUserAsLoggedOut()
-        {
-            var anonymousUser = new ClaimsPrincipal(new ClaimsIdentity());
-            var authState = Task.FromResult(new AuthenticationState(anonymousUser));
-            NotifyAuthenticationStateChanged(authState);
-        }
-
-        private IEnumerable<Claim> ParseClaimsFromJwt(string jwt)
-        {
-            var claims = new List<Claim>();
-            var payload = jwt.Split('.')[1];
-            var jsonBytes = ParseBase64WithoutPadding(payload);
-            var keyValuePairs = JsonSerializer.Deserialize<Dictionary<string, object>>(jsonBytes);
-
-            keyValuePairs.TryGetValue(ClaimTypes.Role, out object roles);
-
-            if (roles != null)
-            {
-                if (roles.ToString().Trim().StartsWith("["))
-                {
-                    var parsedRoles = JsonSerializer.Deserialize<string[]>(roles.ToString());
-
-                    foreach (var parsedRole in parsedRoles)
-                    {
-                        claims.Add(new Claim(ClaimTypes.Role, parsedRole));
-                    }
-                }
-                else
-                {
-                    claims.Add(new Claim(ClaimTypes.Role, roles.ToString()));
-                }
-
-                keyValuePairs.Remove(ClaimTypes.Role);
-            }
-
-            // Add ClaimTypes.Name if available in the JWT payload
-            if (keyValuePairs.TryGetValue("unique_name", out object uniqueName) && uniqueName != null)
-            {
-                claims.Add(new Claim(ClaimTypes.Name, uniqueName.ToString()));
-                keyValuePairs.Remove("unique_name");
-            }
-
-            claims.AddRange(keyValuePairs.Select(kvp => new Claim(kvp.Key, kvp.Value.ToString())));
-            return claims;
-        }
-
-        private byte[] ParseBase64WithoutPadding(string base64)
-        {
-            switch (base64.Length % 4)
-            {
-                case 2: base64 += "=="; break;
-                case 3: base64 += "="; break;
-            }
-            return Convert.FromBase64String(base64);
-        }
-    }
-}
